code3.0/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trade2"
version = "2.0.0"
description = "XAUUSD Automated Quant Research System"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "hmmlearn>=0.3",
    "scikit-learn>=1.3",
    "TA-Lib>=0.4",
    "optuna>=3.0",
    "joblib>=1.3",
    "pyarrow>=14.0",
    "matplotlib>=3.7",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.30", "plotly>=5.18"]
ai = ["anthropic>=0.40"]

[project.scripts]
trade2 = "trade2.app.run_pipeline:main"
full_scheme_search = "trade2.app.full_scheme_search:main"
full_scheme_search_v2 = "trade2.app.full_scheme_search_v2:main"
full_scheme_search_v3 = "trade2.app.full_scheme_search_v3:main"
wf_positive = "trade2.app.wf_positive:main"
test_long_only = "trade2.app.test_long_only:main"
test_3state_4h = "trade2.app.test_3state_4h:main"
trade2-ui = "trade2.app.launch_ui:main"
full_scheme_search_ai = "trade2.app.full_scheme_search_ai:main"
full_scheme_search_3state = "trade2.app.full_scheme_search_3state:main"
tv_research = "trade2.app.tv_research_loop:main"
scalp_research = "trade2.app.scalp_research_loop:main"
trade2-live = "trade2.live.main:main"
trade2-status = "trade2.live.status:run_status"
trade2-replay = "trade2.live.replay:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
